8. Реализация меню за счет шаблона стратегия. [#181780]

+++ src/main/java/ru/job4j/tracker/ConsoleInput.java
@@ -14,7 +14,7 @@ public class ConsoleInput implements Input {
     @Override
     public int askInt(String question) {
         System.out.print(question);
-        if (scanner.hasNextShort()) {
+        if (scanner.hasNextInt()) {
             return scanner.nextInt();
         } else {
             return 7;
